Step 9: delete old app-target RunnerStore files, add RunnerPollerConformances
